# docs/assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IBEF Backend API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/history/{name}": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Get History Metadata",
        "description": "Get metadata for a specific test history.",
        "operationId": "get_history_metadata_api_history__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestMetaData"
                }
              }
            }
          },
          "404": {
            "description": "Test history not found.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Test history 'test_123' not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "history"
        ],
        "summary": "Delete History",
        "description": "Permanently delete a test history.",
        "operationId": "delete_history_api_history__name__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Test history not found.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Test history 'test_123' not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "history"
        ],
        "summary": "Update History Metadata",
        "description": "Update metadata for a test history.",
        "operationId": "update_history_metadata_api_history__name__put",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestMetaData"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Test history not found.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Test history 'test_123' not found"
                }
              }
            }
          },
          "500": {
            "description": "Failed to update metadata.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Failed to update metadata: Permission denied"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/history/{name}/download": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Download History",
        "description": "Download a test history as a ZIP file containing metadata, raw log, and CSV.",
        "operationId": "download_history_api_history__name__download_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Test history not found.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Test history 'test_123' not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/history/{name}/archive": {
      "put": {
        "tags": [
          "history"
        ],
        "summary": "Archive History",
        "description": "Archive a test history (move to archived storage).",
        "operationId": "archive_history_api_history__name__archive_put",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Test history not found.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Test history 'test_123' not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/history/{name}/description": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Get Description",
        "description": "Get the description.md content for a test history.\nReturns the markdown content as plain text.",
        "operationId": "get_description_api_history__name__description_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Description Api History  Name  Description Get"
                }
              }
            }
          },
          "404": {
            "description": "Test history or description not found.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Description not found for test test_123"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "history"
        ],
        "summary": "Update Description",
        "description": "Update the description.md content for a test history.\n    \n    Request body:\n    ```json\n    {\n        \"content\": \"# My Test Description\n\nThis is a markdown file.\"\n    }\n    ```",
        "operationId": "update_description_api_history__name__description_put",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Test history not found.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Test history 'test_123' not found"
                }
              }
            }
          },
          "500": {
            "description": "Failed to update description.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Failed to update description: Permission denied"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sensor/{sensor_id}/data/history": {
      "get": {
        "tags": [
          "sensor"
        ],
        "summary": "Get Sensor Data History",
        "description": "Get historical data points from a sensor (calibrated/processed values).\nWindow is expressed in seconds and must be one of: 30, 60, 120, 300, 600.\nReturns a fixed number of points (based on 30s sampling) evenly spaced across the window.",
        "operationId": "get_sensor_data_history_api_sensor__sensor_id__data_history_get",
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sensor Id"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Window"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PointsList"
                }
              }
            }
          },
          "400": {
            "description": "Invalid sensor_id or window parameter.",
            "content": {
              "application/json": {
                "examples": {
                  "invalid_sensor": {
                    "value": {
                      "detail": "Invalid sensor_id: INVALID. Valid values are: FORCE, DISP_1, DISP_2, DISP_3"
                    }
                  },
                  "invalid_window": {
                    "value": {
                      "detail": "Invalid window: 45. Allowed values are: [30, 60, 120, 300, 600]"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "No test is currently running.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "No test is currently running"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Point": {
        "properties": {
          "time": {
            "type": "number",
            "title": "Time"
          },
          "value": {
            "type": "number",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "time",
          "value"
        ],
        "title": "Point"
      },
      "PointsList": {
        "properties": {
          "list": {
            "items": {
              "$ref": "#/components/schemas/Point"
            },
            "type": "array",
            "title": "List"
          }
        },
        "type": "object",
        "required": [
          "list"
        ],
        "title": "PointsList"
      },
      "TestMetaData": {
        "properties": {
          "test_id": {
            "type": "string",
            "title": "Test Id"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "operator_name": {
            "type": "string",
            "title": "Operator Name"
          },
          "specimen_code": {
            "type": "string",
            "title": "Specimen Code"
          },
          "dim_length": {
            "type": "number",
            "title": "Dim Length",
            "default": 0.0
          },
          "dim_height": {
            "type": "number",
            "title": "Dim Height",
            "default": 0.0
          },
          "dim_width": {
            "type": "number",
            "title": "Dim Width",
            "default": 0.0
          },
          "loading_mode": {
            "type": "string",
            "title": "Loading Mode",
            "default": ""
          },
          "sensor_spacing": {
            "type": "number",
            "title": "Sensor Spacing",
            "default": 0.0
          },
          "ext_support_spacing": {
            "type": "number",
            "title": "Ext Support Spacing",
            "default": 0.0
          },
          "load_point_spacing": {
            "type": "number",
            "title": "Load Point Spacing",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "test_id",
          "date",
          "operator_name",
          "specimen_code"
        ],
        "title": "TestMetaData",
        "description": "Data class representing metadata for a test.\nCompatible with both dataclass operations and Pydantic validation."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}